segregate components
